+++ test/integration/fixtures/__init__.py
@@ -9,8 +9,6 @@
 ######################################################################
 from __future__ import annotations
 
-import os
-
 import pytest
 from .. import get_b2_auth_data