+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+
+### Added
+* Commands `ls` and `rm` now support shell-like patterns for wildcard matching beside of glob-like patterns.
+
 ## [1.23.0] - 2023-08-10
 
 ### Added
@@ -478,8 +482,7 @@ has changed.
 ### Added
 Initial official release of SDK as a separate package (until now it was a part of B2 CLI)
 
-[Unreleased]: https://github.com/Backblaze/b2-sdk-python/compare/v1.23.0...HEAD
-[1.23.0]: https://github.com/Backblaze/b2-sdk-python/compare/v1.22.1...v1.23.0
+[Unreleased]: https://github.com/Backblaze/b2-sdk-python/compare/v1.22.1...HEAD
 [1.22.1]: https://github.com/Backblaze/b2-sdk-python/compare/v1.22.0...v1.22.1
 [1.22.0]: https://github.com/Backblaze/b2-sdk-python/compare/v1.21.0...v1.22.0
 [1.21.0]: https://github.com/Backblaze/b2-sdk-python/compare/v1.20.0...v1.21.0